requirements.txt
attrs
matplotlib
numpy
scipy
